1.格数版本国标修改

+++ src/main/java/com/genersoft/iot/vmp/common/InviteInfo.java
@@ -11,7 +11,7 @@ public class InviteInfo {
 
     private String deviceId;
 
-    private Integer channelId;
+    private Long channelId;
 
     private String stream;
 
@@ -42,7 +42,7 @@ public class InviteInfo {
     private String endTime;
 
 
-    public static InviteInfo getInviteInfo(String deviceId, Integer channelId, String stream, SSRCInfo ssrcInfo, String mediaServerId,
+    public static InviteInfo getInviteInfo(String deviceId, Long channelId, String stream, SSRCInfo ssrcInfo, String mediaServerId,
                                            String receiveIp, Integer receivePort, String streamMode,
                                            InviteSessionType type, InviteSessionStatus status, Boolean record) {
         InviteInfo inviteInfo = new InviteInfo();
