+++ src/main/java/com/genersoft/iot/vmp/gb28181/controller/ChannelFrontEndController.java
@@ -43,7 +43,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "tiltSpeed", description = "垂直速度(0-100)", required = true)
     @Parameter(name = "zoomSpeed", description = "缩放速度(0-100)", required = true)
     @GetMapping("/ptz")
-    public DeferredResult<WVPResult<String>> ptz(Integer channelId, String command, Integer panSpeed, Integer tiltSpeed, Integer zoomSpeed){
+    public DeferredResult<WVPResult<String>> ptz(Long channelId, String command, Integer panSpeed, Integer tiltSpeed, Integer zoomSpeed){
 
         if (log.isDebugEnabled()) {
             log.debug("[通用通道]云台控制 API调用，channelId：{} ，command：{} ，panSpeed：{} ，tiltSpeed：{} ，zoomSpeed：{}",channelId, command, panSpeed, tiltSpeed, zoomSpeed);
@@ -134,7 +134,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "command", description = "控制指令,允许值: in, out, stop", required = true)
     @Parameter(name = "speed", description = "光圈速度(0-100)", required = true)
     @GetMapping("/fi/iris")
-    public DeferredResult<WVPResult<String>> iris(Integer channelId, String command, Integer speed){
+    public DeferredResult<WVPResult<String>> iris(Long channelId, String command, Integer speed){
 
         if (log.isDebugEnabled()) {
             log.debug("[通用通道]光圈控制 API调用，channelId：{} ，command：{} ，speed：{} ",channelId, command, speed);
@@ -188,7 +188,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "command", description = "控制指令,允许值: near, far, stop", required = true)
     @Parameter(name = "speed", description = "聚焦速度(0-100)", required = true)
     @GetMapping("/fi/focus")
-    public DeferredResult<WVPResult<String>> focus(Integer channelId, String command, Integer speed){
+    public DeferredResult<WVPResult<String>> focus(Long channelId, String command, Integer speed){
 
         if (log.isDebugEnabled()) {
             log.debug("[通用通道]聚焦控制 API调用，channelId：{} ，command：{} ，speed：{} ", channelId, command, speed);
@@ -238,7 +238,7 @@ public class ChannelFrontEndController {
     @Operation(summary = "查询预置位", security = @SecurityRequirement(name = JwtUtils.HEADER))
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @GetMapping("/preset/query")
-    public DeferredResult<WVPResult<List<Preset>>> queryPreset(Integer channelId) {
+    public DeferredResult<WVPResult<List<Preset>>> queryPreset(Long channelId) {
         if (log.isDebugEnabled()) {
             log.debug("[通用通道] 预置位查询API调用, {}", channelId);
         }
@@ -266,7 +266,7 @@ public class ChannelFrontEndController {
         return result;
     }
 
-    private DeferredResult<WVPResult<String>> controlPreset(Integer channelId, FrontEndControlCodeForPreset controlCode) {
+    private DeferredResult<WVPResult<String>> controlPreset(Long channelId, FrontEndControlCodeForPreset controlCode) {
         CommonGBChannel channel = channelService.getOne(channelId);
         Assert.notNull(channel, "通道不存在");
 
@@ -295,7 +295,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "presetId", description = "预置位编号", required = true)
     @Parameter(name = "presetName", description = "预置位名称", required = true)
     @GetMapping("/preset/add")
-    public DeferredResult<WVPResult<String>> addPreset(Integer channelId, Integer presetId, String presetName) {
+    public DeferredResult<WVPResult<String>> addPreset(Long channelId, Integer presetId, String presetName) {
         FrontEndControlCodeForPreset controlCode = new FrontEndControlCodeForPreset();
         controlCode.setCode(1);
         controlCode.setPresetId(presetId);
@@ -308,7 +308,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "presetId", description = "预置位编号(1-100)", required = true)
     @GetMapping("/preset/call")
-    public DeferredResult<WVPResult<String>> callPreset(Integer channelId, Integer presetId) {
+    public DeferredResult<WVPResult<String>> callPreset(Long channelId, Integer presetId) {
         FrontEndControlCodeForPreset controlCode = new FrontEndControlCodeForPreset();
         controlCode.setCode(2);
         controlCode.setPresetId(presetId);
@@ -320,7 +320,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "presetId", description = "预置位编号(1-100)", required = true)
     @GetMapping("/preset/delete")
-    public DeferredResult<WVPResult<String>> deletePreset(Integer channelId, Integer presetId) {
+    public DeferredResult<WVPResult<String>> deletePreset(Long channelId, Integer presetId) {
 
         FrontEndControlCodeForPreset controlCode = new FrontEndControlCodeForPreset();
         controlCode.setCode(3);
@@ -329,7 +329,7 @@ public class ChannelFrontEndController {
         return controlPreset(channelId, controlCode);
     }
 
-    private DeferredResult<WVPResult<String>> tourControl(Integer channelId, FrontEndControlCodeForTour controlCode) {
+    private DeferredResult<WVPResult<String>> tourControl(Long channelId, FrontEndControlCodeForTour controlCode) {
         CommonGBChannel channel = channelService.getOne(channelId);
         Assert.notNull(channel, "通道不存在");
 
@@ -357,7 +357,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "tourId", description = "巡航组号", required = true)
     @Parameter(name = "presetId", description = "预置位编号", required = true)
     @GetMapping("/tour/point/add")
-    public DeferredResult<WVPResult<String>> addTourPoint(Integer channelId, Integer tourId, Integer presetId) {
+    public DeferredResult<WVPResult<String>> addTourPoint(Long channelId, Integer tourId, Integer presetId) {
 
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(1);
@@ -372,7 +372,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "tourId", description = "巡航组号(1-100)", required = true)
     @Parameter(name = "presetId", description = "预置位编号(0-100, 为0时删除整个巡航)", required = true)
     @GetMapping("/tour/point/delete")
-    public DeferredResult<WVPResult<String>> deleteCruisePoint(Integer channelId, Integer tourId, Integer presetId) {
+    public DeferredResult<WVPResult<String>> deleteCruisePoint(Long channelId, Integer tourId, Integer presetId) {
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(2);
         controlCode.setPresetId(presetId);
@@ -387,7 +387,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "speed", description = "巡航速度(1-4095)", required = true)
     @Parameter(name = "presetId", description = "预置位编号", required = true)
     @GetMapping("/tour/speed")
-    public DeferredResult<WVPResult<String>> setCruiseSpeed(Integer channelId, Integer tourId, Integer speed, Integer presetId) {
+    public DeferredResult<WVPResult<String>> setCruiseSpeed(Long channelId, Integer tourId, Integer speed, Integer presetId) {
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(3);
         controlCode.setTourSpeed(speed);
@@ -402,7 +402,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "time", description = "巡航停留时间(1-4095)", required = true)
     @Parameter(name = "presetId", description = "预置位编号", required = true)
     @GetMapping("/tour/time")
-    public DeferredResult<WVPResult<String>> setCruiseTime(Integer channelId, Integer tourId, Integer time, Integer presetId) {
+    public DeferredResult<WVPResult<String>> setCruiseTime(Long channelId, Integer tourId, Integer time, Integer presetId) {
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(4);
         controlCode.setTourTime(time);
@@ -415,7 +415,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "tourId", description = "巡航组号)", required = true)
     @GetMapping("/tour/start")
-    public DeferredResult<WVPResult<String>> startCruise(Integer channelId, Integer tourId) {
+    public DeferredResult<WVPResult<String>> startCruise(Long channelId, Integer tourId) {
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(5);
         controlCode.setTourId(tourId);
@@ -426,14 +426,14 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "tourId", description = "巡航组号", required = true)
     @GetMapping("/tour/stop")
-    public DeferredResult<WVPResult<String>> stopCruise(Integer channelId, Integer tourId) {
+    public DeferredResult<WVPResult<String>> stopCruise(Long channelId, Integer tourId) {
         FrontEndControlCodeForTour controlCode = new FrontEndControlCodeForTour();
         controlCode.setCode(6);
         controlCode.setTourId(tourId);
         return tourControl(channelId, controlCode);
     }
 
-    private DeferredResult<WVPResult<String>> scanControl(Integer channelId, FrontEndControlCodeForScan controlCode) {
+    private DeferredResult<WVPResult<String>> scanControl(Long channelId, FrontEndControlCodeForScan controlCode) {
 
         CommonGBChannel channel = channelService.getOne(channelId);
         Assert.notNull(channel, "通道不存在");
@@ -461,7 +461,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "scanId", description = "扫描组号(0-100)", required = true)
     @GetMapping("/scan/start")
-    public DeferredResult<WVPResult<String>> startScan(Integer channelId, Integer scanId) {
+    public DeferredResult<WVPResult<String>> startScan(Long channelId, Integer scanId) {
         FrontEndControlCodeForScan controlCode = new FrontEndControlCodeForScan();
         controlCode.setCode(1);
         controlCode.setScanId(scanId);
@@ -473,7 +473,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "scanId", description = "扫描组号(0-100)", required = true)
     @GetMapping("/scan/stop")
-    public DeferredResult<WVPResult<String>> stopScan(Integer channelId, Integer scanId) {
+    public DeferredResult<WVPResult<String>> stopScan(Long channelId, Integer scanId) {
         FrontEndControlCodeForScan controlCode = new FrontEndControlCodeForScan();
         controlCode.setCode(5);
         controlCode.setScanId(scanId);
@@ -484,7 +484,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "scanId", description = "扫描组号(0-100)", required = true)
     @GetMapping("/scan/set/left")
-    public DeferredResult<WVPResult<String>> setScanLeft(Integer channelId, Integer scanId) {
+    public DeferredResult<WVPResult<String>> setScanLeft(Long channelId, Integer scanId) {
         FrontEndControlCodeForScan controlCode = new FrontEndControlCodeForScan();
         controlCode.setCode(2);
         controlCode.setScanId(scanId);
@@ -495,7 +495,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "scanId", description = "扫描组号(0-100)", required = true)
     @GetMapping("/scan/set/right")
-    public DeferredResult<WVPResult<String>> setScanRight(Integer channelId, Integer scanId) {
+    public DeferredResult<WVPResult<String>> setScanRight(Long channelId, Integer scanId) {
         FrontEndControlCodeForScan controlCode = new FrontEndControlCodeForScan();
         controlCode.setCode(3);
         controlCode.setScanId(scanId);
@@ -508,7 +508,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "scanId", description = "扫描组号(0-100)", required = true)
     @Parameter(name = "speed", description = "自动扫描速度(1-4095)", required = true)
     @GetMapping("/scan/set/speed")
-    public DeferredResult<WVPResult<String>> setScanSpeed(Integer channelId, Integer scanId, Integer speed) {
+    public DeferredResult<WVPResult<String>> setScanSpeed(Long channelId, Integer scanId, Integer speed) {
         FrontEndControlCodeForScan controlCode = new FrontEndControlCodeForScan();
         controlCode.setCode(4);
         controlCode.setScanId(scanId);
@@ -521,7 +521,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "channelId", description = "通道国标编号", required = true)
     @Parameter(name = "command", description = "控制指令,允许值: on, off", required = true)
     @GetMapping("/wiper")
-    public DeferredResult<WVPResult<String>> wiper(Integer channelId, String command){
+    public DeferredResult<WVPResult<String>> wiper(Long channelId, String command){
 
         CommonGBChannel channel = channelService.getOne(channelId);
         Assert.notNull(channel, "通道不存在");
@@ -564,7 +564,7 @@ public class ChannelFrontEndController {
     @Parameter(name = "command", description = "控制指令,允许值: on, off", required = true)
     @Parameter(name = "auxiliaryId", description = "开关编号", required = true)
     @GetMapping("/auxiliary")
-    public DeferredResult<WVPResult<String>> auxiliarySwitch(Integer channelId, String command, Integer auxiliaryId){
+    public DeferredResult<WVPResult<String>> auxiliarySwitch(Long channelId, String command, Integer auxiliaryId){
 
         CommonGBChannel channel = channelService.getOne(channelId);
         Assert.notNull(channel, "通道不存在");
