+++ src/main/java/com/genersoft/iot/vmp/gb28181/controller/bean/ChannelToRegionByGbDeviceParam.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class ChannelToRegionByGbDeviceParam {
-    private List<Integer> deviceIds;
+    private List<Long> deviceIds;
     private String civilCode;
 }
